Mainly reorganization and the eval method for meta_cat
