+++ medcat/utils/ml_utils.py
@@ -34,216 +34,6 @@ def get_batch(ind, batch_size, x, y, cpos, device):
     return x_batch.to(device), y_batch.to(device), c_batch.to(device)
 
 
-def create_batch_piped_data(data, start_ind, end_ind, device, pad_id):
-    max_seq_len = max([len(x[0]) for x in data])
-    print(max_seq_len)
-    x = [x[0][0:max_seq_len] + [pad_id]*max(0, max_seq_len - len(x[0])) for x in data[start_ind:end_ind]]
-    print(x)
-    cpos = [x[1] for x in data[start_ind:end_ind]]
-
-    x = torch.tensor(x, dtype=torch.long).to(device)
-    cpos = torch.tensor(cpos, dtype=torch.long).to(device)
-
-    return x, cpos
-
-
-def process_piped_data(model, data, config):
-    pad_id = config.model['padding_idx']
-    batch_size = config.general['batch_size_eval']
-    device = config.general['device']
-    ignore_cpos = config.model['ignore_cpos']
-
-    model.eval()
-    model.to(device)
-
-    num_batches = math.ceil(len(data) / batch_size)
-    all_logits = []
-    for i in range(num_batches):
-        x, cpos = create_batch_piped_data(data, i*batch_size, (i+1)*batch_size, device=device, pad_id=pad_id)
-        logits = model(x, cpos, ignore_cpos=ignore_cpos)
-        all_logits.append(logits.detach().cpu().numpy())
-
-    logits = np.concatenate(all_logits, axis=0)
-    predictions = np.argmax(logits, axis=1)
-    confidence = np.max(logits, axis=1)
-
-    return predictions, confidence
-
-
-def eval_network(net, data, config, save_dir_path=None):
-    " TODO: Refactor "
-    max_seq_len = config.general['cntx_left'] + config.general['cntx_right'] + 10
-    pad_id = config.model['padding_idx']
-    batch_size = config.general['batch_size_eval']
-    device = config.general['device']
-    ignore_cpos = config.model['ignore_cpos']
-    score_average = config.train['score_average']
-
-
-    # TODO: It is simply wrong to do things this way (whole function)
-    y = np.array([x[0] for x in data])
-    x = [x[1] for x in data]
-    cent = np.array([x[2] for x in data])
-
-    # Pad X and convert to array
-    x = np.array([(sample + [pad_id] * max(0, max_seq_len - len(sample)))[0:max_seq_len] for sample in x])
-
-    x_test = torch.tensor(x, dtype=torch.long)
-    y_test = torch.tensor(y, dtype=torch.long)
-    c_test = torch.tensor(cent, dtype=torch.long)
-
-    device = torch.device(device) # Create a torch device
-    # Move network to the right device
-    net.to(device)
-
-    num_batches_test = int(np.ceil(len(x_test) / batch_size))
-    test_outs = []
-    running_loss_test = []
-    criterion = nn.CrossEntropyLoss() # Set the criterion to Cross Entropy Loss
-
-    for j in range(num_batches_test):
-        net.eval()
-        x_test_batch, y_test_batch, cpos_test_batch = get_batch(ind=j,
-                                                                 batch_size=batch_size,
-                                                                 x=x_test,
-                                                                 y=y_test,
-                                                                 cpos=c_test,
-                                                                 device=device)
-        outputs_test = net(x_test_batch, cpos_test_batch, ignore_cpos=ignore_cpos)
-        test_outs.append(outputs_test.detach().cpu().numpy())
-        running_loss_test.append(criterion(outputs_test, y_test_batch).item())
-
-    test_loss = np.average(running_loss_test)
-
-
-    print("*"*50 + "  Test")
-    print(classification_report(y_test, np.argmax(np.concatenate(test_outs, axis=0), axis=1)))
-    print("Test Loss:  {:5}\n\n".format(test_loss))
-    print("\n\n\n")
-
-    cls_report = classification_report(y_test, np.argmax(np.concatenate(test_outs, axis=0), axis=1), output_dict=True)
-    return cls_report
-
-
-def train_network(net, data, config, save_dir_path=None):
-    max_seq_len = config.general['cntx_left'] + config.general['cntx_right'] + 10
-
-    # Split data
-    y = np.array([x[0] for x in data])
-    x = [x[1] for x in data]
-    cent = np.array([x[2] for x in data])
-
-    # Pad X and convert to array
-    pad_id = config.model['padding_idx']
-    x = np.array([(sample + [pad_id] * max(0, max_seq_len - len(sample)))[0:max_seq_len] for sample in x])
-
-    x_train, x_test, y_train, y_test, c_train, c_test = train_test_split(x, y,
-            cent, test_size=config.train['test_size'])
-
-    x_train = torch.tensor(x_train, dtype=torch.long)
-    y_train = torch.tensor(y_train, dtype=torch.long)
-    c_train = torch.tensor(c_train, dtype=torch.long)
-
-    x_test = torch.tensor(x_test, dtype=torch.long)
-    y_test = torch.tensor(y_test, dtype=torch.long)
-    c_test = torch.tensor(c_test, dtype=torch.long)
-
-    device = torch.device(config.general['device']) # Create a torch device
-    class_weights = config.train['class_weights']
-    if class_weights is not None:
-        class_weights = torch.FloatTensor(class_weights).to(device)
-        criterion = nn.CrossEntropyLoss(weight=class_weights) # Set the criterion to Cross Entropy Loss
-    else:
-        criterion = nn.CrossEntropyLoss() # Set the criterion to Cross Entropy Loss
-    parameters = filter(lambda p: p.requires_grad, net.parameters())
-    optimizer = optim.Adam(parameters, lr=config.train['lr'])
-    net.to(device) # Move the network to device
-
-    batch_size = config.train['batch_size']
-    # Calculate the number of batches given training size len(x_train)
-    num_batches = int(np.ceil(len(x_train) / batch_size))
-    best_f1 = 0
-    best_p = 0
-    best_r = 0
-    best_cls_report = None
-    nepochs = config.train['nepochs']
-    ignore_cpos = config.model['ignore_cpos']
-    for epoch in range(nepochs):
-        running_loss_train = []
-        running_loss_test = []
-
-        # TRAIN PHASE
-        net.train()
-        train_outs = []
-        for i in range(num_batches):
-            x_train_batch, y_train_batch, cpos_train_batch = get_batch(ind=i,
-                                                                       batch_size=batch_size,
-                                                                       x=x_train,
-                                                                       y=y_train,
-                                                                       cpos=c_train,
-                                                                       device=device)
-            optimizer.zero_grad()
-            outputs = net(x_train_batch, cpos_train_batch, ignore_cpos=ignore_cpos)
-            loss = criterion(outputs, y_train_batch)
-            loss.backward()
-            running_loss_train.append(loss.item())
-
-            parameters = filter(lambda p: p.requires_grad, net.parameters())
-            torch.nn.utils.clip_grad_norm_(parameters, 0.25)
-            optimizer.step()
-
-            train_outs.append(outputs.detach().cpu().numpy())
-
-
-        # TEST PHASE
-        num_batches_test = int(np.ceil(len(x_test) / batch_size))
-        test_outs = []
-        for j in range(num_batches_test):
-            net.eval()
-            x_test_batch, y_test_batch, cpos_test_batch = get_batch(ind=j,
-                                                                     batch_size=batch_size,
-                                                                     x=x_test,
-                                                                     y=y_test,
-                                                                     cpos=c_test,
-                                                                     device=device)
-            outputs_test = net(x_test_batch, cpos_test_batch, ignore_cpos=ignore_cpos)
-            test_outs.append(outputs_test.detach().cpu().numpy())
-            running_loss_test.append(criterion(outputs_test, y_test_batch).item())
-
-        train_loss = np.average(running_loss_train)
-        test_loss = np.average(running_loss_test)
-
-        print(f'Epoch: {epoch} ' + "*"*50 + "  Train")
-        print(classification_report(y_train, np.argmax(np.concatenate(train_outs, axis=0), axis=1)))
-        print(f'Epoch: {epoch} ' + "*"*50 + "  Test")
-        print(classification_report(y_test, np.argmax(np.concatenate(test_outs, axis=0), axis=1)))
-        print("Train Loss: {:5}\nTest Loss:  {:5}\n\n".format(train_loss, test_loss))
-        print("\n\n\n")
-        score_average = config.train['score_average']
-        f1 = f1_score(y_test, np.argmax(np.concatenate(test_outs, axis=0), axis=1), average=score_average)
-        precision = precision_score(y_test, np.argmax(np.concatenate(test_outs, axis=0), axis=1), average=score_average)
-        recall = recall_score(y_test, np.argmax(np.concatenate(test_outs, axis=0), axis=1), average=score_average)
-        cls_report = classification_report(y_test, np.argmax(np.concatenate(test_outs, axis=0), axis=1), output_dict=True)
-
-        if f1 > best_f1:
-            print("=" * 50)
-            if config.train['auto_save_model']:
-                path = os.path.join(save_dir_path, 'model.dat')
-                torch.save(net.state_dict(), path)
-                print("Model saved to {} at epoch: {} and f1: {}".format(path, epoch, f1))
-
-            best_f1 = f1
-            best_p = precision
-            best_r = recall
-            best_cls_report = cls_report
-
-            # Print some stats
-            print(confusion_matrix(y_test, np.argmax(np.concatenate(test_outs, axis=0), axis=1)))
-            print("\n\n")
-
-    return (best_f1, best_p, best_r, best_cls_report)
-
-
 def load_hf_tokenizer(tokenizer_name):
     try:
         from transformers import AutoTokenizer
